Properly blend internal and external textures
Rather than having a fixed flag behavior, or game type detection, we do check if the attached texture for the draw call is an external texture.
If so, we do use a blending mode that does consider the alpha channel for colors too. If not, we use the classical expected blending inherited from the game engine.
This finally fixes once and for all the fancy transparency behavior, adding a smart layer logic, as well as removing any drawbacks seen as of today.