SFX: This layer has now a concept of re-usable channels and dedicated channels
At this stage, you can specify how many channels are re-usable, but now which ones ( so if you say 5, it's going to be 1-5 ).
You can also specify how many channels you want in total, and what is going to be played on any channel that is > re-usable number, is going to be considered dedicated.
This will allow to defer the unload through an explicit call named unloadSFX.