recalcInternals fix for framebuffer size
This is intended to fix an issue of inconsistent behavior at certain resolutions.  In addition, it should make upscaling the internal resolution from the original game resolution consistent across the board.  This includes changing the default read and FFNx.toml values of internal_resolution_scale to 0 for auto-fit calculation as well as updating the Internal Resolution Scale heading's description on the toml file.