recalcInternals fix and new default
This includes both a fix to how the recalcInternals function works as well as the new default value of 0 for the settings files.  Had to make an additional change to ensure the original internal_resolution_scale is not modified so that when resize events happen, the scaling factor can be recalculated correctly.  Reverted to using long variable types for scaleW, scaleH and the new scale_factor to avoid an issue with the value round() returns not converting to uint8_t as we were expecting.  Also simplified the code slightly at Julian's suggestion by assigning internal_resolution_factor to scale_factor at declaration, meaning the else statement was unnecessary.