60FPS: Fix limits in turn movement opcode (#347)
This fixes the limitation in TURNGEN and TURN opcode in about 8 cases if running double the speed.
It uses external rotation steps index and number of steps when it a special flag is set, otherwise the legacy code is still implemented and works.

Also refactor some FIELD 60 FPS fix functions.