Voice: Fix an edge case for voice file selection filenames
Before the notation X_Y where X was the dialog id and Y was the option, was used to identify an option file,
although now the very same notation could be used also where X is the window id and Y is the dialog id.
In order to fix this clash, prefix the new window detection filenames with 'w'.

Before:
-> 2_1: window_id = 2, dialog_id = 1

After:
-> w2_1: window_id = 2, dialog_id = 1