FF7: Fix OK simulation logic
By default, OK button should be sent only until the window is not being closed.
If it is closing it means the OK button has been already sent and therefore it doesn't need to be sent again.

Fix this behavior therefore by disabling the 'is_voice_acting' flag and ensuring an edge case where:
1) Window is still opened
2) An audio played but is now finished
3) An Opcode is triggering the window to close
Under this condition a double OK would be sent