60FPS: Implement smooth FIELD background movement
Change the update mechanism related to following the player in FIELD mode by moving the background and the 3d models.
All positions of the 4 layers of the background are computed using floating point variables. This is also done for the 3d model world coordinates.
This removes all the wobbling background and 3d models. To avoid changing the original game, this smoothness in FIELD movement is applied only when using FULL 30 FPS or FULL 60 FPS mode.