Renderer: Fix advanced lighting handles
Since the original conception of this code, it was expecting to have an invalid BGFX handle so on next usage the texture would be not used by BGFX.
Although on the other hand the entire FFNx code was expecting 0 when the texture did not exist. This created an interesting weird behavior where BGFX though nothing to do here, but the entire rest of the code considered it an external texture ( avoiding to load the vanilla ones when needed ).