Mod: Fix the override_mod_path layered logic lookup for animated textures
Before it was no more looking on the second pass ( mod_path layer ) because the animated textures always set the key as 0, so when the code is re-executed it will shortcircuit instead of looking up again.
Now it will continue to do the same, but only on the last layer ( mod_path ). So we correctly lookup twice, if the override_mod_path is specified.