Audio: Fix SFX loop case asking to play more than once
In FF8 there might be cases where the engine is asking to play the same SFX over and over,
even though the SFX is meant to loop. Probably because the looping effect is implemented in the engine itself.
Since we handle it ourselves, just skip the play action and keep looping the same effect.